% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-analysis.R
\docType{methods}
\name{calc_fold_change,GenePanel-method}
\alias{calc_fold_change,GenePanel-method}
\title{Calculate Fold Change}
\usage{
\S4method{calc_fold_change}{GenePanel}(object)
}
\arguments{
\item{object}{A GenePanel object.}
}
\value{
A dataframe of results with Log2FC and P-values.
}
\description{
Performs differential expression analysis using Welch's t-test.
}
