% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GenePanel-class}
\alias{GenePanel-class}
\alias{exprs,GenePanel-method}
\alias{meta,GenePanel-method}
\alias{cohort,GenePanel-method}
\title{GenePanel Class}
\usage{
exprs(object)

\S4method{exprs}{GenePanel}(object)

\S4method{meta}{GenePanel}(object)

\S4method{cohort}{GenePanel}(object)
}
\description{
An S4 class to represent a targeted gene expression panel and clinical data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{exprs(GenePanel)}: Get the expression matrix

\item \code{meta(GenePanel)}: Get the clinical metadata

\item \code{cohort(GenePanel)}: Get the cancer type label

}}
\section{Slots}{

\describe{
\item{\code{expression_data}}{A numeric matrix (Genes x Patients).}

\item{\code{patient_metadata}}{A data.frame of clinical info.}

\item{\code{cancer_type}}{A character string describing the cohort.}
}}

