% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\docType{methods}
\name{plot_volcano,GenePanel-method}
\alias{plot_volcano,GenePanel-method}
\title{Volcano Plot}
\usage{
\S4method{plot_volcano}{GenePanel}(object)
}
\arguments{
\item{object}{A GenePanel object.}
}
\value{
A ggplot object.
}
\description{
Generates a volcano plot for differential expression results.
}
