% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-risk.R
\docType{methods}
\name{predict_risk}
\alias{predict_risk}
\title{Predict Risk}
\usage{
predict_risk(object, gene, direction = "low_risk_high_expr")
}
\arguments{
\item{object}{A GenePanel object.}

\item{gene}{The gene symbol to use.}

\item{direction}{'low_risk_high_expr' or 'high_risk_high_expr'.}
}
\value{
A factor of risk labels.
}
\description{
Stratifies patients based on biomarker expression.
}
